[squash] Further progress ...
