--- packages/bpk-docs/src/layouts/NeoSideNav/SideNavLayout.scss
@@ -1,75 +1,0 @@
-/*
- * Backpack - Skyscanner's Design System
- *
- * Copyright 2018 Skyscanner Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@import '~bpk-mixins/index';
-
-.bpkdocs-side-nav-layout {
-  &__wrapper {
-    display: flex;
-    flex-direction: row;
-  }
-
-  &__logo {
-    height: $bpk-spacing-xs * 8;
-
-    // HACK: This is to align the 'B' of 'Backpack' with the left edge
-    // of the content.
-    margin-left: $bpk-one-pixel-rem * 2;
-
-    @include bpk-rtl {
-      margin-left: 0;
-    }
-  }
-
-  &__mobile-nav {
-    display: none;
-    padding-top: $bpk-spacing-sm;
-
-    @include bpk-breakpoint-tablet {
-      display: block;
-    }
-  }
-
-  &__sidebar {
-    display: flex;
-    min-width: $bpk-spacing-base * 20;
-    padding: $bpk-spacing-lg * 2;
-    flex-direction: column;
-    align-items: flex-start;
-    background-color: $bpk-color-gray-900;
-    color: $bpk-color-white;
-
-    @include bpk-breakpoint-tablet {
-      display: none;
-    }
-
-    &-content {
-      width: 100%;
-      margin-top: $bpk-spacing-lg * 3;
-    }
-  }
-
-  &__content {
-    display: flex;
-    padding: $bpk-spacing-sm;
-  }
-
-  &__section-list {
-    margin-bottom: $bpk-spacing-md;
-  }
-}
