+++ packages/bpk-docs/src/layouts/NeoSideNavLayout/SectionsList.scss
@@ -35,6 +35,7 @@ $bpk-heading-button-padding: $bpk-spacing-md;
     background-color: transparent;
     color: $bpk-color-white;
     text-align: left;
+    white-space: nowrap;
     cursor: pointer;
 
     @include bpk-text;
@@ -47,6 +48,7 @@ $bpk-heading-button-padding: $bpk-spacing-md;
   }
 
   &__heading-button-arrow {
+    margin-left: $bpk-spacing-sm;
     fill: currentColor;
   }
 
@@ -74,7 +76,11 @@ $bpk-heading-button-padding: $bpk-spacing-md;
 
   &__list-item {
     display: block;
+    width: 100%;
     padding: $bpk-spacing-md 0;
+    border: none;
+    background: none;
+    text-align: left;
 
     @include bpk-border-bottom-sm($bpk-color-gray-700);
     @include nav-link;
@@ -86,6 +92,7 @@ $bpk-heading-button-padding: $bpk-spacing-md;
     @include bpk-rtl {
       margin-right: $bpk-first-column-size;
       margin-left: 0;
+      text-align: right;
     }
   }
 }
